refactor: port & adaptor 계층 분리
