+++ src/main/java/org/iimsa/hub_service/hub/infrastructure/HubCoordinateAdapter.java
@@ -2,17 +2,17 @@ package org.iimsa.hub_service.hub.infrastructure;
 
 import lombok.RequiredArgsConstructor;
 import org.iimsa.hub_service.hub.domain.repository.HubRepository;
-import org.iimsa.hub_service.hubroute.domain.service.HubCoordinatePort;
+import org.iimsa.hub_service.hubroute.application.HubCoordinatePort;
 import org.springframework.stereotype.Component;
 
 import java.util.Optional;
 import java.util.UUID;
 
 /**
- * HubCoordinatePort 구현체
+ * HubCoordinatePort 구현체 (어댑터)
  *
- * <p>HubRoute 도메인이 Hub의 위도/경도를 조회할 수 있도록 Hub JPA Repository를 위임합니다.
- * Hub 도메인 infrastructure에 위치하여 HubRepository에 직접 접근합니다.
+ * <p>HubRoute application이 Hub의 위도/경도를 조회할 수 있도록
+ * Hub 도메인의 {@link HubRepository}를 위임합니다.
  */
 @Component
 @RequiredArgsConstructor
